+++ cms/djangoapps/contentstore/__init__.py
@@ -1,0 +1,2 @@
+""" module init will register signal handlers """
+import contentstore.signals
